+++ src/InjectionContainer/InjectionContainerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Yii\View\InjectionContainer;
+
+interface InjectionContainerInterface
+{
+    public function get(string $id): object;
+}