+++ src/ViewRenderer.php
@@ -19,6 +19,8 @@ use Yiisoft\View\ViewContextInterface;
 use Yiisoft\View\WebView;
 use Yiisoft\Yii\View\Exception\InvalidLinkTagException;
 use Yiisoft\Yii\View\Exception\InvalidMetaTagException;
+use Yiisoft\Yii\View\InjectionContainer\InjectionContainerInterface;
+use Yiisoft\Yii\View\InjectionContainer\StubInjectionContainer;
 
 use function array_key_exists;
 use function array_merge;
@@ -48,13 +50,22 @@ final class ViewRenderer implements ViewContextInterface
     private ?string $locale = null;
 
     /**
+     * @var object[]|null
+     */
+    private ?array $preparedInjections = null;
+
+    private InjectionContainerInterface $injectionContainer;
+
+    /**
      * @param DataResponseFactoryInterface $responseFactory The data response factory instance.
      * @param Aliases $aliases The aliases instance.
      * @param WebView $view The web view instance.
      * @param string|null $viewPath The full path to the directory of views or its alias.
      * @param string|null $layout The layout name (e.g. "layout/main") to be applied to views.
      * If null, the layout will not be applied.
-     * @param object[] $injections The injection instances.
+     * @param array $injections The injection instances or class names.
+     *
+     * @psalm-param array<object|string> $injections
      */
     public function __construct(
         private DataResponseFactoryInterface $responseFactory,
@@ -62,8 +73,10 @@ final class ViewRenderer implements ViewContextInterface
         private WebView $view,
         ?string $viewPath = null,
         private ?string $layout = null,
-        private array $injections = []
+        private array $injections = [],
+        ?InjectionContainerInterface $injectionContainer = null,
     ) {
+        $this->injectionContainer = $injectionContainer ?? new StubInjectionContainer();
         $this->setViewPath($viewPath);
     }
 
@@ -241,24 +254,24 @@ final class ViewRenderer implements ViewContextInterface
     /**
      * Return a new instance with the appended specified injections.
      *
-     * @param object ...$injections The injection instances.
+     * @param object|string ...$injections The injection instances or class names.
      */
-    public function withAddedInjections(object ...$injections): self
+    public function withAddedInjections(object|string ...$injections): self
     {
         $new = clone $this;
-        $new->injections = array_merge($this->injections, $injections);
+        $new->setInjections(array_merge($this->injections, $injections));
         return $new;
     }
 
     /**
      * Returns a new instance with the specified injections.
      *
-     * @param object ...$injections The injection instances.
+     * @param object|string ...$injections The injection instances or class names.
      */
-    public function withInjections(object ...$injections): self
+    public function withInjections(object|string ...$injections): self
     {
         $new = clone $this;
-        $new->injections = $injections;
+        $new->setInjections($injections);
         return $new;
     }
 
@@ -403,7 +416,7 @@ final class ViewRenderer implements ViewContextInterface
     private function getInjections(?string $layout, string $injectionInterface): array
     {
         $result = [];
-        foreach ($this->injections as $injection) {
+        foreach ($this->getPreparedInjections() as $injection) {
             if ($injection instanceof $injectionInterface) {
                 $result[] = $injection;
                 continue;
@@ -575,5 +588,30 @@ final class ViewRenderer implements ViewContextInterface
     private function setViewPath(?string $path): void
     {
         $this->viewPath = $path === null ? null : rtrim($path, '/');
+    }
+
+    private function getPreparedInjections(): array
+    {
+        if ($this->preparedInjections !== null) {
+            return $this->preparedInjections;
+        }
+
+        $this->preparedInjections = [];
+        foreach ($this->injections as $injection) {
+            $this->preparedInjections[] = is_object($injection)
+                ? $injection
+                : $this->injectionContainer->get($injection);
+        }
+
+        return $this->preparedInjections;
+    }
+
+    /**
+     * @psalm-param array<object|string> $injections
+     */
+    private function setInjections(array $injections): void
+    {
+        $this->injections = $injections;
+        $this->preparedInjections = null;
     }
 }