+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enh #79: Add debug collector for yiisoft/yii-debug (@xepozz)
 - Bug #82: Fixed find for layout file due to compatibility with `yiisoft/view` (@rustamwin)
 - Enh #99: Make `viewPath` in `ViewRenderer` constructor optional (@vjik)
+- Enh #107: Implement lazy loading for injections (@vjik)
 - New #102: Add layout specific injections (@vjik) 
 
 ## 6.0.0 February 16, 2023